Agregando nulls al Post

+++ RIAE3.1/Models/Boletas.cs
@@ -5,6 +5,6 @@
         public int IdBoleta { get; set; }
         public int IdRegistro { get; set; }
         public int IdParametro { get; set; }
-        public float ImporteUnitarioClasificador { get; set; }
+        public decimal ImporteUnitarioClasificador { get; set; }
     }
 }
